conditional and access restrictions (#1240)
* added timedomain

* moved dow logic to parser.  added logic for restriction conditional.

* added time domain tests.

* init version of td.

* added MONTH enum and new restriction types timedallowed and timedenied.

* added logic to parse conditions.  Will return multiple if they are found.

* added modes to access restrictions.  Don't default to truck anymore.

* removed hour on and off and min on and off logic.  Time info is now stored in the 64 bit value time_domain.

* created a struct for from and to graphids so that we can use the spare bits to store time domain information for complex restrictions.

* use new accessor graphid functions.

* removed hard coded truck mode.

* added logic to process conditional restrictions.  Removed logic for hour on and off and min on and off.

* updated

* updated to use reutil.

* updated year

* updated to support values of uint 64 for restrictions.

* do not allow routing where there is only an allowed time or denied time for now.  TODO: add date/time logic.

* added is_restricted tests and fixed conditional restriction tests.

* added is_restricted function.  Removed year.  Fixed range issue for dow range Sa-Su.

* added getters for time info on the complex restriction.

* added setters for time info on the complex restriction.

* added logic to set the time on the complex restriction.

* clean up

* added logic to set the time on the complex restriction.

* added debug info.

* removed year from vars and functions.

* added comment.

* added is_restricted function.

* Updates to simplify ComplexRestriction. ComplexRestrictionBuilder now
derives from ComplexRestriction. Rework to avoid copying and creating
a vector of Vias, rather use pointers to the structure and the via list.

* clean up and added counts.

* clean up and added via check.

* Change ComplexRestriction to a fixed size structure (no need for a
constructor anymore) and add code to read the via list which follows
the structure to DynamicCost - the only place where restrictions
are read.

* updated begin_day to begin_day_dow and end_day to end_day_dow.

* updated begin_day to begin_day_dow and end_day to end_day_dow.  value now has an accessor

* updated comment

* added set methods.  added range/validity checks to set methods where fields can overflow.  made the daterange protected.

* removed todo.  range checks are done at the data processing time.

* added timezone_diff for depart_at and arrive_by routing.  some code clean up and added seconds_to_date function.

* updated start out from seconds_since_epoch and added logic to call the tz diff function so that we can add the tz offset to the localtime.

* localtime is now uint64_t.

* added tests for tz diff.

* updated comments

+++ valhalla/thor/timedep.h
@@ -54,7 +54,7 @@ class TimeDepForward : public AStarPathAlgorithm {
    * @param  pred_idx     Predecessor index into the EdgeLabel list.
    * @param  from_transition True if this method is called from a transition
    *                         edge.
-   * @param  localtime    Current local time.
+   * @param  localtime    Current local time.  Seconds since epoch
    * @param  dest         Location information of the destination.
    * @param  best_path    Best path found so far. Includes the index into
    *                      EdgeLabels and the cost.
@@ -62,7 +62,7 @@ class TimeDepForward : public AStarPathAlgorithm {
   void ExpandForward(baldr::GraphReader& graphreader,
                      const baldr::GraphId& node, const sif::EdgeLabel& pred,
                      const uint32_t pred_idx, const bool from_transition,
-                     uint32_t localtime, const odin::Location& dest,
+                     uint64_t localtime, const odin::Location& dest,
                      std::pair<int32_t, float>& best_path);
 
   /**
@@ -137,7 +137,7 @@ class TimeDepReverse : public AStarPathAlgorithm {
    * @param  opp_pred_edge Opposing predecessor directed edge.
    * @param  from_transition True if this method is called from a transition
    *                         edge.
-   * @param  localtime    Current local time.
+   * @param  localtime    Current local time.  Seconds since epoch
    * @param  dest         Location information of the destination.
    * @param  best_path    Best path found so far. Includes the index into
    *                      EdgeLabels and the cost.
@@ -147,7 +147,7 @@ class TimeDepReverse : public AStarPathAlgorithm {
                      const uint32_t pred_idx,
                      const baldr::DirectedEdge* opp_pred_edge,
                      const bool from_transition,
-                     uint32_t localtime, const odin::Location& dest,
+                     uint64_t localtime, const odin::Location& dest,
                      std::pair<int32_t, float>& best_path);
 
   /**
